[risk] Rename xml to orexml
