+++ projects/ores.risk/orexml/currency_serialiser.hpp
@@ -1,6 +1,6 @@
 /* -*- mode: c++; tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 4 -*-
  *
- * Copyright (C) 2025 Marco Craveiro <marco.craveiro@gmail.com>
+ * Copyright (C) 2024 Marco Craveiro <marco.craveiro@gmail.com>
  *
  * This program is free software; you can redistribute it and/or modify it under
  * the terms of the GNU General Public License as published by the Free Software
@@ -17,22 +17,22 @@
  * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  *
  */
-#ifndef ORES_RISK_XML_CURRENCY_CONFIG_SERIALISER_HPP
-#define ORES_RISK_XML_CURRENCY_CONFIG_SERIALISER_HPP
+#ifndef ORES_RISK_OREXML_CURRENCY_SERIALISER_HPP
+#define ORES_RISK_OREXML_CURRENCY_SERIALISER_HPP
 
 #if defined(_MSC_VER) && (_MSC_VER >= 1200)
 #pragma once
 #endif
 
-#include <string>
-#include "ores.risk/domain/currency_config.hpp"
+#include <rapidxml-ns/rapidxml_ns.hpp>
+#include "ores.risk/domain/currency.hpp"
 
-namespace ores::risk::xml {
+namespace ores::risk::orexml {
 
-class currency_config_serialiser {
+class currency_serialiser {
 public:
-    std::string serialise(const domain::currency_config& cfg);
-    domain::currency_config deserialise(std::string s);
+    void serialise(rapidxml_ns::xml_node<>& parent, const domain::currency& cfg);
+    domain::currency deserialise(rapidxml_ns::xml_node<>& node);
 };
 
 }
